+++ Arduino/libraries/Edt-devices/DMXLedSpot.h
@@ -62,7 +62,7 @@ class LedSpot : public Slave
 	}
 
   public:
-	void initialize(int address, int maximumBrightness, int minimumBrightness)
+	void initialize(uint16_t address, uint8_t maximumBrightness, uint8_t minimumBrightness)
 	{
 		_address = address;
 		_maximumBrightness = maximumBrightness;
